+++ stories/constants/colorComponent.js
